requirements.txt
Jinja2==3.1.6
numpy==2.3.5
openai==2.13.0
Pillow==12.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
mss>=9.0.0
pyautogui>=0.9.54
pyperclip>=1.8.0
